코드 리뷰 반영하여 코드 수정
1. [Footer.js] 시맨틱한 태그 작성
2. variables.css 생성하여 자주 쓰이는 CSS 전역 변수 선언
3. 열거형 enum 사용해 상품 타입을 정형화된 값으로 관리
4. 리액트 영역의 코드와 스타일을 담는 코드 파일 분리
