+++ src/components/Footer/Footer.js
@@ -1,0 +1,15 @@
+import React from "react";
+import * as S from './style'
+
+function Footer() {
+  return (
+    <S.FooterContainer>
+      <S.Text>
+        개인정보 처리방침 | 이용 약관<br />
+        All rights reserved @ Codestates
+      </S.Text>
+    </S.FooterContainer>
+  )
+}
+
+export default Footer;
