+++ src/components/Footer/style.js
@@ -1,0 +1,16 @@
+import styled from "styled-components";
+import '../../variables.css'
+
+export const FooterContainer = styled.div`
+position: relative;
+height: 58px;
+border-top: 1px solid var(--black-trans-100);
+`;
+
+export const Text = styled.div`
+  text-align: center;
+  font-size: 12px;
+  color: var(--footer-gray);
+  margin-top: 11px;
+  margin-bottom: 11px;
+  `;
